anomaly-detection-app/dashboard/backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
psycopg2-binary>=2.9.1
pydantic>=1.8.2
python-dateutil>=2.8.2
kafka-python>=2.0.2
requests>=2.26.0
sse-starlette>=0.10.3
asyncpg>=0.25.0
python-dotenv>=0.19.0
fastapi-utils>=0.2.1
aiohttp>=3.8.1
slack-sdk>=3.21.0